[ADD] pos_sale_picking_inherit_order
For standard, when pickings are created for a transferred POS Order from a
Sales Order, picking contact becomes order contact, ignoring when SO shipping
address is different, and customer reference is lost.

With this addon such pickings are created ensuring thar picking contact is
the same that backend should be set, that is SO shipping address, and the same
with customer reference as well.